+++ src/ui.rs
@@ -560,14 +560,14 @@ impl<E, M> Ui<E, M> {
       // TODO: Ideally we would want to go without the recursion stuff we
       //       have. This may not be possible (efficiently) with safe
       //       Rust, though. Not sure.
-      let bbox = widget.render(self, renderer, bbox);
+      let inner_bbox = widget.render(self, renderer, bbox);
 
       // We start rendering with the widget with the lowest z-index,
       // i.e., the one the furthest in the background.
       for child_id in self.children(idx).rev() {
         let child_idx = self.validate(*child_id);
         let child = self.lookup(child_idx);
-        self.render_all(child_idx, child, renderer, bbox)
+        self.render_all(child_idx, child, renderer, inner_bbox)
       }
 
       let () = widget.render_done(self, renderer, bbox);